requirements.txt
argcomplete==1.9.2
asgiref==3.4.1
base58==0.2.4
boto3==1.18.44
botocore==1.21.44
certifi==2021.5.30
cfn-flip==1.0.4
charset-normalizer==2.0.6
click==8.0.1
Django==3.2.7
django-filter==2.4.0
django-s3-storage==0.13.4
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
django-cors-headers==3.8.0
Pillow==8.3.2
docutils==0.17.1
durationpy==0.5
future==0.16.0
hjson==3.0.1
idna==3.2
jmespath==0.9.3
kappa==0.6.0
lambda-packages==0.19.0
Markdown==3.3.4
pep517==0.11.0
pip-tools==6.2.0
placebo==0.9.0
PyJWT==2.1.0
python-dateutil==2.6.1
python-slugify==1.2.4
pytz==2021.1
PyYAML==3.12
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
tomli==1.2.1
tqdm==4.19.1
troposphere==3.0.3
Unidecode==1.3.2
urllib3==1.26.6
Werkzeug==0.12
wsgi-request-logger==0.4.6
zappa==0.53.0